pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "LangGraph ReAct agent"
authors = [
    {name = "Viktor Szabo",email = "vszabo@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage (>=2.0.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-cloud-aiplatform (>=1.60.0,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
